https://app.asana.com/0/41180108506105/42983347910395
NOTE: There was no primary key on ERM diagram. When I reached the update and deletes area it didn't make sense to me not to have a primary key on this class. Yes, there were two foreign keys what I think is known as a joint class (don't really know what that is but have heard of it in theory). product_permissions is a many to many relations and only having foreign keys when deleting and updating just doesn't work or maybe it does...let me know if I have to edit code OR if keeping pass the word onto the others...for the sql tables at least...would there be any other place for this to be updated??? All the other ERs have primary keys. Is one is the odd ball.